+++ src/ext/conf/BrowserDetect.js
@@ -1,3 +1,5 @@
+import browser from "vuex-webextensions/dist/browser";
+
 if (process.env.CHANNEL !== 'stable') {
   console.info('Loaded BrowserDetect');
 }
@@ -9,6 +11,9 @@ const BrowserDetect = {
   edge: process.env.BROWSER === 'edge',
   processEnvBrowser: process.env.BROWSER,
   processEnvChannel: process.env.CHANNEL,
+  isEdgeUA: () => /Edg\/(\.?[0-9]*)*$/.test(window.navigator.userAgent),
+  getBrowserObj: () => { return process.env.BROWSER === 'firefox' ? browser : chrome; },
+  getURL: (url) => { console.log('getting file:', url); console.log(process.env.BROWSER === 'firefox' ? browser.runtime.getURL(url) : chrome.runtime.getURL(url)); return process.env.BROWSER === 'firefox' ? browser.runtime.getURL(url) : chrome.runtime.getURL(url); },
 } 
 
 if (process.env.CHANNEL !== 'stable') {